refactor: move numberFormatValidator func and tests in utility files
