+++ src/util/util.spec.ts
@@ -1,5 +1,11 @@
-import { getURILastElement, isError, isSubstringOf, endsZeroDotted } from './util'
 import * as O from 'fp-ts/Option'
+import {
+  getURILastElement,
+  isError,
+  isSubstringOf,
+  endsZeroDotted,
+  numberFormatValidator
+} from './util'
 
 type Data = {
   arg: string
@@ -88,5 +94,52 @@ describe('endsZeroDotted', () => {
 
   it('returns false for empty string', () => {
     expect(endsZeroDotted('')).toBe(false)
+  })
+})
+
+describe('numberFormatValidator', () => {
+  const thousandSeparator = ' '
+  const decimalSeparator = '.'
+  const decimalLimit = 2
+  const validate = numberFormatValidator(thousandSeparator, decimalSeparator, decimalLimit)
+
+  it('returns none when the string contains several occurences of a decimal separator', () => {
+    expect(validate('123.123.123')).toEqual(O.none)
+    expect(validate('123..')).toEqual(O.none)
+    expect(validate('..123')).toEqual(O.none)
+  })
+
+  it('returns none when the string starts with a leading dash (-)', () => {
+    expect(validate('-123')).toEqual(O.none)
+  })
+
+  it('returns none when the string contains non-alphanumeric characters that does not belong to separators', () => {
+    expect(validate('123%')).toEqual(O.none)
+    expect(validate('Hello')).toEqual(O.none)
+  })
+
+  it('returns none when the string contains more decimal than the limit', () => {
+    expect(validate('12.345')).toEqual(O.none)
+  })
+
+  it('returns none when the string contains thousands separators at the wrong position', () => {
+    expect(validate('1 2345')).toEqual(O.none)
+    expect(validate('123 45')).toEqual(O.none)
+  })
+
+  it('returns none when the string ends with a thousand separator', () => {
+    expect(validate('123 ')).toEqual(O.none)
+  })
+
+  it('returns some when the string ends with a decimal separator', () => {
+    expect(validate('123.')).toEqual(O.some('123.'))
+  })
+
+  it('returns some when the string format is valid', () => {
+    expect(validate('')).toEqual(O.some(''))
+    expect(validate('123.12')).toEqual(O.some('123.12'))
+    expect(validate('12 345')).toEqual(O.some('12 345'))
+    expect(validate('12 345 678')).toEqual(O.some('12 345 678'))
+    expect(validate('12 345 678.2')).toEqual(O.some('12 345 678.2'))
   })
 })
